+++ react/src/index.tsx
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom/client'
 import './style.css'
 import './common.css'
 
-import { Router } from './navigation/navigator'
+import { Router } from './navigation/routers'
 
 function loadPage(): void {
     const root = ReactDOM.createRoot(document.getElementById('root')!)
